requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.2
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic[email]==2.5.0
bcrypt==4.1.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyJWT==2.8.1
slowapi==0.1.9
httpx==0.25.0
builtwith==1.3.5
dnspython==2.4.2
whois==0.9.7
requests==2.31.0
reportlab==4.0.7
Pillow==10.1.0
